# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/pets/owner/{owner_id}": {
            "get": {
                "description": "Get a pet list given the owner ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Get pets of owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "owner id to get pets",
                        "name": "owner_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset of the results",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit of the results ",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SearchResponse-model_Pet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
        "/pets/pet": {
            "post": {
                "description": "Create a pet for a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Creates a Pet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "description": "pet info",
                        "name": "pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Pet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Pet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
        "/pets/pet/{id}": {
            "get": {
                "description": "Get pet info given a pet ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Get a Pet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "id of the pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Pet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit pet info given a pet ID and pet info needed to update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Edit a Pet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "id of the pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "pet info",
                        "name": "pet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Pet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Pet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a pet given a pet ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pet"
                ],
                "summary": "Delete a Pet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "request from telegram",
                        "name": "X-Telegram-App",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chat id of the telegram user",
                        "name": "X-Telegram-Id",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "id of the pet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "controller.Pet": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "2013-05-23"
                },
                "name": {
                    "type": "string",
                    "example": "Raaida"
                },
                "owner_id": {
                    "type": "string",
                    "example": "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AnimalType"
                        }
                    ],
                    "example": "dog"
                }
            }
        },
...
        "model.AnimalType": {
            "type": "string",
            "enum": [
                "dog",
                "cat",
                "bird",
                "hamster"
            ],
            "x-enum-varnames": [
                "Dog",
                "Cat",
                "Bird",
                "Hamster"
            ]
        },
...
        "model.Paging": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.Pet": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "register_date": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.AnimalType"
                }
            }
        },
        "model.SearchResponse-model_Pet": {
            "type": "object",
            "properties": {
                "paging": {
                    "$ref": "#/definitions/model.Paging"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Pet"
                    }
                }
            }
        },
...
    }
}